fix: move special condition tooltips logic to Output.tsx
