+++ client/src/app/components/Tooltip.scss
@@ -3,7 +3,6 @@
 .tooltip {
     position: relative;
     display: flex;
-    align-items: center;
 
     .wrap {
         visibility: hidden;
