+++ client/src/app/components/stardust/AssociatedOutputsTable.scss
@@ -95,11 +95,6 @@
                     margin: 0px;
                 }
             }
-
-            .icon {
-                color: $gray-10!important;
-                font-size: 20px!important;
-            }
         }
 
         .output--label {
