+++ client/src/app/components/stardust/OutputState.tsx
@@ -7,4 +7,9 @@ export interface OutputState {
      * Toggle balance between raw and formatted amount.
      */
     isFormattedBalance: boolean;
+
+    /**
+     * Show special unlock condition.
+     */
+    isSpecialCondition: boolean;
 }
